♻️ Adding custom hook to improve reusability
since I got several components using the same logic, I just add it as a custom hook
