+++ cypress.config.ts
@@ -1,8 +1,9 @@
-import { defineConfig } from "cypress";
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import { defineConfig } from 'cypress';
 
 export default defineConfig({
   e2e: {
-    setupNodeEvents(on, config) {
+    setupNodeEvents(_on, _config) {
       // implement node event listeners here
     },
   },
